further auth tweaks and update seeder data

+++ WorldTravel/WorldTravel.Application/Countries/Commands/DeleteCountry/DeleteCountryCommandHandler.cs
@@ -1,12 +1,14 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using WorldTravel.Domain.Constants;
 using WorldTravel.Domain.Entities;
 using WorldTravel.Domain.Exceptions;
+using WorldTravel.Domain.Interface;
 using WorldTravel.Domain.Repositories;
 
 namespace WorldTravel.Application.Countries.Commands.DeleteCountry;
 
-public class DeleteCountryCommandHandler(ILogger<DeleteCountryCommandHandler> logger, ICountriesRepository countriesRepository) 
+public class DeleteCountryCommandHandler(ILogger<DeleteCountryCommandHandler> logger, ICountriesRepository countriesRepository, ICountryAuthorizationService countryAuthorizationService) 
     : IRequestHandler<DeleteCountryCommand>
 {
     public async Task Handle(DeleteCountryCommand request, CancellationToken cancellationToken)
@@ -14,6 +16,13 @@ public class DeleteCountryCommandHandler(ILogger<DeleteCountryCommandHandler> lo
         logger.LogInformation($"Deleting country with Id: {request.Id}");
         var country = await countriesRepository.GetByIdAsync(request.Id) 
             ?? throw new NotFoundException(nameof(Country), request.Id);
+
+        if (!countryAuthorizationService.Authorize(country, ResourceOperation.Delete))
+        {
+            logger.LogWarning($"User is not authorized to delete country with Id: {request.Id}");
+            throw new ForbidException();
+        }
+
         await countriesRepository.DeleteAsync(country);
     }
 }
